requirements.txt
flask
flet
gtts
openai
python-dotenv
better_profanity
